Add more packet types

+++ capybara/src/network/client/native.rs
@@ -81,7 +81,7 @@ impl WebSocketClient {
                                 }
                                 Packet::Pong { timestamp } => {
                                     let now = match SystemTime::now().duration_since(SystemTime::UNIX_EPOCH) {
-                                        Ok(now) => now.as_millis(),
+                                        Ok(now) => now.as_millis() as u64,
                                         Err(err) => error_continue!("Failed to obtain current time ({})", err),
                                     };
